+++ src/ONTraC/bin/ONTraC_analysis.py
@@ -2,6 +2,7 @@ import os
 import sys
 from optparse import OptionGroup, OptionParser, Values
 
+from ..analysis.niche_net_constr import embedding_adjust_visualization
 from ..analysis.cell_type import cell_type_visualization
 from ..analysis.data import AnaData
 from ..analysis.niche_cluster import niche_cluster_visualization
@@ -25,16 +26,19 @@ def analysis_pipeline(options: Values) -> None:
     # 0. load data class
     ana_data = AnaData(options)
 
-    # part 1: train loss
+    # part 1: niche network construction
+    embedding_adjust_visualization(ana_data=ana_data)
+
+    # part 2: train loss
     train_loss_visualiztion(ana_data=ana_data)
 
-    # part 2: spatial based output
+    # part 3: spatial based output
     spatial_visualization(ana_data=ana_data)
 
-    # part 3: niche cluster
+    # part 4: niche cluster
     niche_cluster_visualization(ana_data=ana_data)
 
-    # part 4: cell type based output
+    # part 5: cell type based output
     cell_type_visualization(ana_data=ana_data)
 
 
@@ -51,6 +55,24 @@ def add_suppress_group(optparser: OptionParser) -> None:
                      action='store_true',
                      default=False,
                      help='Suppress the niche cluster loadings visualization.')
+    optparser.add_option_group(group)
+
+
+def add_embedding_adjust_group(optparser: OptionParser) -> None:
+    group = OptionGroup(optparser, 'Embedding adjust options')
+    group.add_option(
+        '--embedding-adjust',
+        dest='embedding_adjust',
+        action='store_true',
+        default=False,
+        help=
+        'Adjust the cell type coding according to embeddings. Default is False. At least two (Embedding_1 and Embedding_2) should be in the original data if embedding_adjust is True.'
+    )
+    group.add_option(
+        '--sigma',
+        dest='sigma',
+        type='float',
+        help='Sigma for the exponential function. Default is the mean of the distances between the cell type pairs.')
     optparser.add_option_group(group)
 
 
@@ -82,6 +104,7 @@ def prepare_optparser() -> OptionParser:
                          help='Plot each sample separately.')
     add_IO_options_group(optparser=optparser, io_options=IO_OPTIONS)
     add_suppress_group(optparser)
+    add_embedding_adjust_group(optparser)
     return optparser